requirements.txt
appdirs==1.4.4
asgiref==3.8.1
baron==0.10.1
dj-database-url==2.2.0
dj-email-url==1.0.6
Django==5.0.6
django-cache-url==3.4.5
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
environs==11.0.0
marshmallow==3.21.3
packaging==24.1
PyJWT==2.8.0
python-dotenv==1.0.1
redbaron==0.9.2
rply==0.7.8
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
